+++ R/pat_aggregateOutlierCounts.R
@@ -50,6 +50,9 @@ pat_aggregateOutlierCounts <-
     if ( pat_isEmpty(pat) )
       stop("Required parameter 'pat' has no data.") 
     
+    # Remove any duplicate data records
+    pat <- pat_distinct(pat)
+
     # ----- Convert period to seconds --------------------------------------------
     
     periodParts <- strsplit(period, " ", fixed = TRUE)[[1]]