+++ R/pat_externalFit.R
@@ -68,6 +68,9 @@ pat_externalFit <- function(
   if ( pat_isEmpty(pat) )
     stop("Parameter 'pat' has no data.")
   
+  # Remove any duplicate data records
+  pat <- pat_distinct(pat)
+  
   # For easier access
   meta <- pat$meta
   data <- pat$data