+++ R/pat_internalFit.R
@@ -58,6 +58,9 @@ pat_internalFit <- function(
   if ( pat_isEmpty(pat) )
     stop("Parameter 'pat' has no data.")
   
+  # Remove any duplicate data records
+  pat <- pat_distinct(pat)
+  
   # For easier access
   meta <- pat$meta
   data <- pat$data