+++ R/pat_loadLatest.R
@@ -58,6 +58,9 @@ pat_loadLatest <- function(
     stop(paste0("Data file could not be loaded: ", filepath), call.=FALSE)
   }
   
+  # Remove any duplicate data records
+  pat <- pat_distinct(pat)
+  
   return(pat)
   
 }