+++ R/pat_monitorComparison.R
@@ -79,6 +79,9 @@ pat_monitorComparison <- function(
   if ( pat_isEmpty(pat) )
     stop("Parameter 'pat' has no data.")
 
+  # Remove any duplicate data records
+  pat <- pat_distinct(pat)
+  
   # ----- Assemble data --------------------------------------------------------
   
   if ( replaceOutliers )