+++ R/pat_multiplot.R
@@ -81,6 +81,9 @@ pat_multiplot <- function(
       stop("Parameter 'pat' has no data.")
   }
   
+  # Remove any duplicate data records
+  pat <- pat_distinct(pat)
+  
   # ----- Reduce large datasets by sampling ------------------------------------
   
   if ( !is.null(pat) && !is.null(sampleSize) ) {