+++ R/pas_staticMap.R
@@ -49,6 +49,7 @@
 #'
 #' @examples
 #' \donttest{
+#' setArchiveBaseUrl("http://smoke.mazamascience.com/data/PurpleAir")
 #' LA_basin <- 
 #'   pas_load() %>% 
 #'   pas_filterArea(-118.5, -117.5, 33.5, 34.5)