+++ tests/testthat/test-pas_load.R
@@ -3,12 +3,12 @@ context("test-pas_load")
 test_that("baseUrl Input validation works", {
   expect_error(downloadParseSynopticData(baseUrl = "https://www.purpleair.com/DUMMY"))
   expect_error(pas_createNew(baseUrl = "https://www.purpleair.com/DUMMY"))
-  expect_error(pas_load(baseUrl = "https://www.mazamascience.com/DUMMY"))
 })
 
 test_that("loads correct class", {
   skip_on_cran()
   skip_on_travis()
+  setArchiveBaseUrl("http://smoke.mazamascience.com/data/PurpleAir")
   expect_true(pas_isPas(pas_load()))
   # expect_true(pas_isPas(pas_createNew()))
 })